+++ src/i18n.ts
@@ -165,6 +165,9 @@ export enum I18N {
   welcomeCreateWallet,
   welcomeLedgerWallet,
   welcomeRestoreWallet,
+  backupFinishCongratulation,
+  backupFinishSuccess,
+  backupFinishFinish
 }
 
 export function getText(key: I18N, params?: Record<string, string>): string {
@@ -366,4 +369,7 @@ const en: Record<I18N, string> = {
   [I18N.welcomeCreateWallet]: 'Create Wallet',
   [I18N.welcomeLedgerWallet]: 'Connect',
   [I18N.welcomeRestoreWallet]: 'I already have a wallet',
+  [I18N.backupFinishCongratulation]: 'Congratulations!',
+  [I18N.backupFinishSuccess]: "You've successfully protected your wallet.!",
+  [I18N.backupFinishFinish]: 'Finish',
 };
